update:清楚无效的目录
